feat: move data processing to separate service
