+++ rs-lang-app/src/app/core/services/process-statistics.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ProcessStatisticsService } from './process-statistics.service';
+
+describe('ProcessStatisticsService', () => {
+  let service: ProcessStatisticsService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(ProcessStatisticsService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
