+++ rs-lang-app/src/app/sprint/components/sprint/sprint.component.html
@@ -86,9 +86,7 @@
         </div>
       </div>
       <div class="stats-buttons">
-        <button
-          class="button-green"
-          (click)="resetGame(); shuffleWords(wordsArray)">
+        <button class="button-green" (click)="resetGame(); restartGame()">
           play again
         </button>
         <button class="button-orange" (click)="resetGame(); goBack()">
