+++ rs-lang-app/src/app/sprint/components/sprint/sprint.component.ts
@@ -37,6 +37,7 @@ import { QueryService } from 'src/app/core/service/query.service';
 import { HttpClient } from '@angular/common/http';
 import { ActivatedRoute, NavigationEnd, Router } from '@angular/router';
 import { GameLevelTransferService } from 'src/app/core/services/game-level-transfer.service';
+import { ProcessStatisticsService } from 'src/app/core/services/process-statistics.service';
 
 @Component({
   selector: 'app-sprint',
@@ -88,7 +89,8 @@ export class SprintComponent implements OnInit {
     private httpService: HttpService,
     private queryService: QueryService,
     private _location: Location,
-    private LevelPage: GameLevelTransferService
+    private LevelPage: GameLevelTransferService,
+    private processStatisticsService: ProcessStatisticsService
   ) {
     this.isAuth = this.userService.isRegistered();
     this.isAuth
@@ -241,7 +243,12 @@ export class SprintComponent implements OnInit {
             ? Math.round((this.successWords / this.totalWords) * 100)
             : 0;
         if (this.totalWords > 0 && this.isAuth) {
-          this.processStatistics();
+          // this.processStatistics();
+          this.processStatisticsService.serviceData = {
+            gameStats: this.gameStats,
+            longestCombo: this.longestCombo,
+          };
+          this.processStatisticsService.processStatistics();
         }
       }
     }, 100);
@@ -303,7 +310,7 @@ export class SprintComponent implements OnInit {
   resetGame() {
     if (this.timerID) clearInterval(this.timerID);
     this.isGameEnded = false;
-    this.isGameStarted = true;
+    this.isGameStarted = false;
     this.gameStats = [];
     this.timer = SPRINT_TIMER * 10;
     this.gameScore = 0;
@@ -311,6 +318,18 @@ export class SprintComponent implements OnInit {
     this.comboBonus = 0;
     this.longestCombo = 0;
     this.wordsCounter = 0;
+    this.loadingProgress = 0;
+    this.progress = '';
+  }
+
+  restartGame() {
+    if (this.LevelPage.gamePageLevel.length) {
+      this.currentLevel = this.LevelPage.gamePageLevel[0];
+      this.loadWords(this.LevelPage.gamePageLevel[1]);
+    }
+    if (!this.LevelPage.gamePageLevel.length) {
+      this.loadWords();
+    }
   }
 
   goBack() {
@@ -333,201 +352,193 @@ export class SprintComponent implements OnInit {
     });
   }
 
-  putUserStatistics(options: GameStatistics) {
-    // console.log('Previous user Statistics: ', this.userGamesStats);
-    this.queryService.setUserStatistics(options).subscribe({
-      next: (data: any) => {
-        // console.log('New user Statistics: ', data);
-        this.userGamesStats = data;
-      },
-    });
-  }
+  // putUserStatistics(options: GameStatistics) {
+  //   this.queryService.setUserStatistics(options).subscribe({
+  //     next: (data: any) => {
+  //       this.userGamesStats = data;
+  //     },
+  //   });
+  // }
 
-  processUserWord(word: IWord, wordGameStats: ISprintStats) {
-    // console.log('•••USER WORD••• Old data: ', word);
-    const optionsWord: IWordsData = {
-      difficulty: Difficulty.Learned,
-      optional: {
-        rightGuessesInRow: 0,
-      },
-    };
-    if (wordGameStats.success) {
-      if (
-        (word.optional?.rightGuessesInRow === 2 &&
-          word.difficulty !== Difficulty.Hard) ||
-        (word.optional?.rightGuessesInRow &&
-          word.optional?.rightGuessesInRow >= 4 &&
-          word.difficulty === Difficulty.Hard) ||
-        (word.difficulty === Difficulty.Easy &&
-          word.optional?.dateEasy === undefined)
-      ) {
-        optionsWord.difficulty = Difficulty.Easy;
-        optionsWord.optional.dateEasy = Date.now();
-        this.removeWordFromArray(word);
-      }
-      if (
-        (word.optional?.rightGuessesInRow &&
-          word.optional?.rightGuessesInRow > 2 &&
-          word.difficulty !== Difficulty.Hard) ||
-        word.difficulty === Difficulty.Easy
-      ) {
-        optionsWord.difficulty = Difficulty.Easy;
-        this.removeWordFromArray(word);
-      }
-      if (word.optional?.rightGuessesInRow === undefined) {
-        optionsWord.optional.rightGuessesInRow = 1;
-      }
-      if (word.optional?.rightGuessesInRow !== undefined) {
-        optionsWord.optional.rightGuessesInRow =
-          word.optional.rightGuessesInRow + 1;
-      }
-      if (word.optional?.dateFirstTime === undefined)
-        optionsWord.optional.dateFirstTime = Date.now();
-      if (word.optional?.dateFirstTime !== undefined)
-        optionsWord.optional.dateFirstTime = word.optional?.dateFirstTime;
-    }
-    if (!wordGameStats.success) {
-      if (word.difficulty === Difficulty.Easy) {
-        optionsWord.difficulty = Difficulty.Learned;
-        optionsWord.optional.rightGuessesInRow = 0;
-        delete optionsWord.optional.dateEasy;
-      }
-      if (word.optional?.dateFirstTime !== undefined) {
-        optionsWord.optional.dateFirstTime = word.optional?.dateFirstTime;
-      }
-      if (word.optional?.dateFirstTime === undefined) {
-        optionsWord.optional.dateFirstTime = Date.now();
-      }
-    }
-    const locationWord =
-      QueryParams.register +
-      SLASH +
-      this.userId +
-      QueryParams.words +
-      SLASH +
-      word.wordId;
-    // this.httpService.putData(locationWord, optionsWord);
-    const response = this.httpService.putData(locationWord, optionsWord);
-    response.subscribe({
-      next: data => console.log('•••USER WORD••• New data: ', data),
-    });
-  }
+  // processUserWord(word: IWord, wordGameStats: ISprintStats) {
+  //   const optionsWord: IWordsData = {
+  //     difficulty: Difficulty.Learned,
+  //     optional: {
+  //       rightGuessesInRow: 0,
+  //     },
+  //   };
+  //   if (wordGameStats.success) {
+  //     if (
+  //       (word.optional?.rightGuessesInRow === 2 &&
+  //         word.difficulty !== Difficulty.Hard) ||
+  //       (word.optional?.rightGuessesInRow &&
+  //         word.optional?.rightGuessesInRow >= 4 &&
+  //         word.difficulty === Difficulty.Hard) ||
+  //       (word.difficulty === Difficulty.Easy &&
+  //         word.optional?.dateEasy === undefined)
+  //     ) {
+  //       optionsWord.difficulty = Difficulty.Easy;
+  //       optionsWord.optional.dateEasy = Date.now();
+  //       this.removeWordFromArray(word);
+  //     }
+  //     if (
+  //       (word.optional?.rightGuessesInRow &&
+  //         word.optional?.rightGuessesInRow > 2 &&
+  //         word.difficulty !== Difficulty.Hard) ||
+  //       word.difficulty === Difficulty.Easy
+  //     ) {
+  //       optionsWord.difficulty = Difficulty.Easy;
+  //       this.removeWordFromArray(word);
+  //     }
+  //     if (word.optional?.rightGuessesInRow === undefined) {
+  //       optionsWord.optional.rightGuessesInRow = 1;
+  //     }
+  //     if (word.optional?.rightGuessesInRow !== undefined) {
+  //       optionsWord.optional.rightGuessesInRow =
+  //         word.optional.rightGuessesInRow + 1;
+  //     }
+  //     if (word.optional?.dateFirstTime === undefined)
+  //       optionsWord.optional.dateFirstTime = Date.now();
+  //     if (word.optional?.dateFirstTime !== undefined)
+  //       optionsWord.optional.dateFirstTime = word.optional?.dateFirstTime;
+  //   }
+  //   if (!wordGameStats.success) {
+  //     if (word.difficulty === Difficulty.Easy) {
+  //       optionsWord.difficulty = Difficulty.Learned;
+  //       optionsWord.optional.rightGuessesInRow = 0;
+  //       delete optionsWord.optional.dateEasy;
+  //     }
+  //     if (word.optional?.dateFirstTime !== undefined) {
+  //       optionsWord.optional.dateFirstTime = word.optional?.dateFirstTime;
+  //     }
+  //     if (word.optional?.dateFirstTime === undefined) {
+  //       optionsWord.optional.dateFirstTime = Date.now();
+  //     }
+  //   }
+  //   const locationWord =
+  //     QueryParams.register +
+  //     SLASH +
+  //     this.userId +
+  //     QueryParams.words +
+  //     SLASH +
+  //     word.wordId;
+  //   const response = this.httpService.putData(locationWord, optionsWord);
+  //   response.subscribe({
+  //     next: data => console.log('•••USER WORD••• New data: ', data),
+  //   });
+  // }
 
-  processNotUserWord(wordId: string, wordGameStats: ISprintStats) {
-    // console.log('>>>NOT USER WORD<<< Old data: ', wordGameStats);
-    const optionsWord: IWordsData = {
-      difficulty: Difficulty.Learned,
-      optional: {
-        rightGuessesInRow: wordGameStats.success ? 1 : 0,
-        dateFirstTime: Date.now(),
-      },
-    };
-    const locationWord =
-      QueryParams.register +
-      SLASH +
-      this.userId +
-      QueryParams.words +
-      SLASH +
-      wordId;
-    // this.httpService.postData(locationWord, optionsWord);
-    const response = this.httpService.postData(locationWord, optionsWord);
-    response.subscribe({
-      next: data => console.log('>>>NOT USER WORD<<< New data: ', data),
-    });
-  }
+  // processNotUserWord(wordId: string, wordGameStats: ISprintStats) {
+  //   const optionsWord: IWordsData = {
+  //     difficulty: Difficulty.Learned,
+  //     optional: {
+  //       rightGuessesInRow: wordGameStats.success ? 1 : 0,
+  //       dateFirstTime: Date.now(),
+  //     },
+  //   };
+  //   const locationWord =
+  //     QueryParams.register +
+  //     SLASH +
+  //     this.userId +
+  //     QueryParams.words +
+  //     SLASH +
+  //     wordId;
+  //   const response = this.httpService.postData(locationWord, optionsWord);
+  //   response.subscribe({
+  //     next: data => console.log('>>>NOT USER WORD<<< New data: ', data),
+  //   });
+  // }
 
-  processStatistics() {
-    this.getUserStatistics();
-    const optionsStat: GameStatistics = {
-      learnedWords: 0,
-      optional: {
-        sprint: {
-          today: {
-            attempts: 0,
-            success: 0,
-            rightGuessesInRow: 0,
-            date: 0,
-          },
-          allTime: {
-            attempts: 0,
-            success: 0,
-            rightGuessesInRow: 0,
-          },
-        },
-      },
-    };
-    // console.log('UserGameStats: ', this.userGamesStats);
-    const optionsSprintStats = optionsStat.optional.sprint;
-    const dayNow = new Date().getDate().toString();
-    const monthNow = new Date().getMonth().toString();
-    const yearNow = new Date().getFullYear().toString();
-    if (
-      this.userGamesStats?.optional !== undefined &&
-      this.userGamesStats?.optional.sprint !== undefined
-    ) {
-      const userSprintStats = this.userGamesStats?.optional.sprint;
-      const statsDay = new Date(<number>userSprintStats.today.date)
-        .getDate()
-        .toString();
-      const statsMonth = new Date(<number>userSprintStats.today.date)
-        .getMonth()
-        .toString();
-      const statsYear = new Date(<number>userSprintStats.today.date)
-        .getFullYear()
-        .toString();
-      optionsSprintStats!.allTime.attempts =
-        userSprintStats.allTime.attempts + this.gameStats.length;
-      optionsSprintStats!.allTime.success =
-        userSprintStats.allTime.success +
-        this.gameStats.filter((el: ISprintStats) => el.success).length;
-      optionsSprintStats!.allTime.rightGuessesInRow =
-        userSprintStats.allTime.rightGuessesInRow >= this.longestCombo
-          ? userSprintStats.allTime.rightGuessesInRow
-          : this.longestCombo;
-      if (
-        dayNow === statsDay &&
-        monthNow === statsMonth &&
-        yearNow === statsYear
-      ) {
-        optionsSprintStats!.today.attempts =
-          userSprintStats.today.attempts + this.gameStats.length;
-        optionsSprintStats!.today.success =
-          userSprintStats.today.success +
-          this.gameStats.filter((el: ISprintStats) => el.success).length;
-        optionsSprintStats!.today.rightGuessesInRow =
-          userSprintStats.today.rightGuessesInRow >= this.longestCombo
-            ? userSprintStats.today.rightGuessesInRow
-            : this.longestCombo;
-        optionsSprintStats!.today.date = userSprintStats.today.date;
-      }
-      if (
-        dayNow !== statsDay ||
-        monthNow !== statsMonth ||
-        yearNow !== statsYear
-      ) {
-        optionsSprintStats!.today.attempts = 0;
-        optionsSprintStats!.today.success = 0;
-        optionsSprintStats!.today.rightGuessesInRow = 0;
-        optionsSprintStats!.today.date = new Date().getTime();
-      }
-    }
-    this.putUserStatistics(optionsStat);
-    this.getUserWords();
-    const userWordIds = this.userWords.map((el: IWord) => el.wordId);
-    // console.log(this.userWords);
-    this.gameStats.forEach((el: ISprintStats) => {
-      userWordIds.includes(el.id)
-        ? this.processUserWord(
-            <IWord>this.userWords.find((word: IWord) => word.wordId === el.id),
-            el
-          )
-        : this.processNotUserWord(el.id, el);
-    });
-  }
+  // processStatistics() {
+  //   this.getUserStatistics();
+  //   const optionsStat: GameStatistics = {
+  //     learnedWords: 0,
+  //     optional: {
+  //       sprint: {
+  //         today: {
+  //           attempts: 0,
+  //           success: 0,
+  //           rightGuessesInRow: 0,
+  //           date: 0,
+  //         },
+  //         allTime: {
+  //           attempts: 0,
+  //           success: 0,
+  //           rightGuessesInRow: 0,
+  //         },
+  //       },
+  //     },
+  //   };
+  //   const optionsSprintStats = optionsStat.optional.sprint;
+  //   const dayNow = new Date().getDate().toString();
+  //   const monthNow = new Date().getMonth().toString();
+  //   const yearNow = new Date().getFullYear().toString();
+  //   if (
+  //     this.userGamesStats?.optional !== undefined &&
+  //     this.userGamesStats?.optional.sprint !== undefined
+  //   ) {
+  //     const userSprintStats = this.userGamesStats?.optional.sprint;
+  //     const statsDay = new Date(<number>userSprintStats.today.date)
+  //       .getDate()
+  //       .toString();
+  //     const statsMonth = new Date(<number>userSprintStats.today.date)
+  //       .getMonth()
+  //       .toString();
+  //     const statsYear = new Date(<number>userSprintStats.today.date)
+  //       .getFullYear()
+  //       .toString();
+  //     optionsSprintStats!.allTime.attempts =
+  //       userSprintStats.allTime.attempts + this.gameStats.length;
+  //     optionsSprintStats!.allTime.success =
+  //       userSprintStats.allTime.success +
+  //       this.gameStats.filter((el: ISprintStats) => el.success).length;
+  //     optionsSprintStats!.allTime.rightGuessesInRow =
+  //       userSprintStats.allTime.rightGuessesInRow >= this.longestCombo
+  //         ? userSprintStats.allTime.rightGuessesInRow
+  //         : this.longestCombo;
+  //     if (
+  //       dayNow === statsDay &&
+  //       monthNow === statsMonth &&
+  //       yearNow === statsYear
+  //     ) {
+  //       optionsSprintStats!.today.attempts =
+  //         userSprintStats.today.attempts + this.gameStats.length;
+  //       optionsSprintStats!.today.success =
+  //         userSprintStats.today.success +
+  //         this.gameStats.filter((el: ISprintStats) => el.success).length;
+  //       optionsSprintStats!.today.rightGuessesInRow =
+  //         userSprintStats.today.rightGuessesInRow >= this.longestCombo
+  //           ? userSprintStats.today.rightGuessesInRow
+  //           : this.longestCombo;
+  //       optionsSprintStats!.today.date = userSprintStats.today.date;
+  //     }
+  //     if (
+  //       dayNow !== statsDay ||
+  //       monthNow !== statsMonth ||
+  //       yearNow !== statsYear
+  //     ) {
+  //       optionsSprintStats!.today.attempts = 0;
+  //       optionsSprintStats!.today.success = 0;
+  //       optionsSprintStats!.today.rightGuessesInRow = 0;
+  //       optionsSprintStats!.today.date = new Date().getTime();
+  //     }
+  //   }
+  //   this.putUserStatistics(optionsStat);
+  //   this.getUserWords();
+  //   const userWordIds = this.userWords.map((el: IWord) => el.wordId);
+  //   this.gameStats.forEach((el: ISprintStats) => {
+  //     userWordIds.includes(el.id)
+  //       ? this.processUserWord(
+  //           <IWord>this.userWords.find((word: IWord) => word.wordId === el.id),
+  //           el
+  //         )
+  //       : this.processNotUserWord(el.id, el);
+  //   });
+  // }
 
-  removeWordFromArray(word: IWord) {
-    this.wordsArray = this.wordsArray.filter(
-      (el: IWordCard) => el.id !== word.wordId
-    );
-  }
+  // removeWordFromArray(word: IWord) {
+  //   this.wordsArray = this.wordsArray.filter(
+  //     (el: IWordCard) => el.id !== word.wordId
+  //   );
+  // }
 }
